Rename parquet producer
